Fix task 2 compilation errror

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -4,7 +4,7 @@
  * @n: the bit
  * @index: the index
  */
-int get_bit(unsigned long int n, unsigned long int index)
+int get_bit(unsigned long int n, unsigned int index)
 {
 	unsigned long int i;
